+++ CHANGELOG.md
@@ -7,6 +7,12 @@
   `Stream<List<String>>`. This passes a stream to stderr without modifying it,
   which is useful for debugging.
 
+* Add a `Script.mapLines` constructor that returns a script that maps stdin
+  lines according to a `String Function(String)`.
+
+* Add support for passing a `String Function(String)` to `Script.pipeline` and
+  `Script.operator |` to map stdin lines.
+
 * Fold stack frames for a few more packages used internally.
 
 ## 0.2.2